requirements.txt
minio>=7.2.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
pyarrow>=14.0.0
pyyaml>=6.0.1
pytest>=8.0.0
pytest-mock>=3.12.0
